# update SQLite PRAGMAs for faster DB